requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
langchain
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
pypdf
python-dotenv
tiktoken
python-multipart
langchain-huggingface